+++ app/src/main/AndroidManifest.xml
@@ -78,7 +78,7 @@
             android:name="asset_statements"
             android:resource="@string/asset_statements" />
 
-        <receiver android:name="de.tum.informatics.www1.artemis.native_app.feature.push.service.impl.notification_manager.ReplyReceiver" />
-        <receiver android:name="de.tum.informatics.www1.artemis.native_app.feature.push.service.impl.notification_manager.DeleteNotificationReceiver" />
+        <receiver android:name="de.tum.informatics.www1.artemis.native_app.feature.push.service.impl.notification_manager.reply.ReplyReceiver" />
+        <receiver android:name="de.tum.informatics.www1.artemis.native_app.feature.push.service.impl.notification_manager.delete.DeleteNotificationReceiver" />
     </application>
 </manifest>
